+++ firebase-firestore/src/androidTest/java/com/google/firebase/firestore/QueryTest.java
@@ -47,7 +47,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.Semaphore;
 import org.junit.After;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -1203,13 +1202,6 @@ public class QueryTest {
         "doc4",
         "doc5");
 
-    // with one inequality: a>2 || b==1.
-    checkOnlineAndOfflineResultsMatch(
-        collection.where(Filter.or(Filter.greaterThan("a", 2), Filter.equalTo("b", 1))),
-        "doc5",
-        "doc2",
-        "doc3");
-
     // (a==1 && b==0) || (a==3 && b==2)
     checkOnlineAndOfflineResultsMatch(
         collection.where(
@@ -1233,6 +1225,35 @@ public class QueryTest {
             Filter.and(
                 Filter.or(Filter.equalTo("a", 2), Filter.equalTo("b", 2)),
                 Filter.or(Filter.equalTo("a", 3), Filter.equalTo("b", 3)))),
+        "doc3");
+
+    // Test with limits without orderBy (the __name__ ordering is the tie breaker).
+    checkOnlineAndOfflineResultsMatch(
+        collection.where(Filter.or(Filter.equalTo("a", 2), Filter.equalTo("b", 1))).limit(1),
+        "doc2");
+  }
+
+  @Test
+  public void testOrQueriesWithCompositeIndexes() {
+    assumeTrue(
+        "Skip this test if running against production because it results in a "
+            + "'missing index' error. The Firestore Emulator, however, does serve these "
+            + " queries.",
+        isRunningAgainstEmulator());
+    Map<String, Map<String, Object>> testDocs =
+        map(
+            "doc1", map("a", 1, "b", 0),
+            "doc2", map("a", 2, "b", 1),
+            "doc3", map("a", 3, "b", 2),
+            "doc4", map("a", 1, "b", 3),
+            "doc5", map("a", 1, "b", 1));
+    CollectionReference collection = testCollectionWithDocs(testDocs);
+
+    // with one inequality: a>2 || b==1.
+    checkOnlineAndOfflineResultsMatch(
+        collection.where(Filter.or(Filter.greaterThan("a", 2), Filter.equalTo("b", 1))),
+        "doc5",
+        "doc2",
         "doc3");
 
     // Test with limits (implicit order by ASC): (a==1) || (b > 0) LIMIT 2
@@ -1266,17 +1287,10 @@ public class QueryTest {
             .limitToLast(1)
             .orderBy("a"),
         "doc2");
-
-    // Test with limits without orderBy (the __name__ ordering is the tie breaker).
-    checkOnlineAndOfflineResultsMatch(
-        collection.where(Filter.or(Filter.equalTo("a", 2), Filter.equalTo("b", 1))).limit(1),
-        "doc2");
   }
 
-  // TODO(orquery): Enable this test when prod supports OR queries.
-  @Ignore
   @Test
-  public void testOrQueriesWithInAndNotIn() {
+  public void testOrQueriesWithIn() {
     Map<String, Map<String, Object>> testDocs =
         map(
             "doc1", map("a", 1, "b", 0),
@@ -1293,6 +1307,24 @@ public class QueryTest {
         "doc3",
         "doc4",
         "doc6");
+  }
+
+  @Test
+  public void testOrQueriesWithNotIn() {
+    assumeTrue(
+        "Skip this test if running against production because it results in a "
+            + "'missing index' error. The Firestore Emulator, however, does serve these "
+            + " queries",
+        isRunningAgainstEmulator());
+    Map<String, Map<String, Object>> testDocs =
+        map(
+            "doc1", map("a", 1, "b", 0),
+            "doc2", map("b", 1),
+            "doc3", map("a", 3, "b", 2),
+            "doc4", map("a", 1, "b", 3),
+            "doc5", map("a", 1),
+            "doc6", map("a", 2));
+    CollectionReference collection = testCollectionWithDocs(testDocs);
 
     // a==2 || b not-in [2,3]
     // Has implicit orderBy b.
@@ -1302,8 +1334,6 @@ public class QueryTest {
         "doc2");
   }
 
-  // TODO(orquery): Enable this test when prod supports OR queries.
-  @Ignore
   @Test
   public void testOrQueriesWithArrayMembership() {
     Map<String, Map<String, Object>> testDocs =
@@ -1332,9 +1362,12 @@ public class QueryTest {
         "doc6");
   }
 
-  @Ignore
   @Test
   public void testMultipleInOps() {
+    // TODO(orquery): Enable this test against production when possible.
+    assumeTrue(
+        "Skip this test if running against production because it's not yet supported.",
+        isRunningAgainstEmulator());
     Map<String, Map<String, Object>> testDocs =
         map(
             "doc1", map("a", 1, "b", 0),
@@ -1347,63 +1380,24 @@ public class QueryTest {
 
     // Two IN operations on different fields with disjunction.
     Query query1 =
-        collection
-            .where(Filter.or(Filter.inArray("a", asList(2, 3)), Filter.inArray("b", asList(0, 2))))
-            .orderBy("a");
-    checkOnlineAndOfflineResultsMatch(query1, "doc1", "doc6", "doc3");
-
-    // Two IN operations on different fields with conjunction.
-    Query query2 =
-        collection
-            .where(Filter.and(Filter.inArray("a", asList(2, 3)), Filter.inArray("b", asList(0, 2))))
-            .orderBy("a");
-    checkOnlineAndOfflineResultsMatch(query2, "doc3");
-
-    // Two IN operations on the same field.
-    // a IN [1,2,3] && a IN [0,1,4] should result in "a==1".
-    Query query3 =
         collection.where(
-            Filter.and(Filter.inArray("a", asList(1, 2, 3)), Filter.inArray("a", asList(0, 1, 4))));
-    checkOnlineAndOfflineResultsMatch(query3, "doc1", "doc4", "doc5");
+            Filter.or(Filter.inArray("a", asList(2, 3)), Filter.inArray("b", asList(0, 2))));
+    checkOnlineAndOfflineResultsMatch(query1, "doc1", "doc3", "doc6");
 
-    // a IN [2,3] && a IN [0,1,4] is never true and so the result should be an empty set.
-    Query query4 =
-        collection.where(
-            Filter.and(Filter.inArray("a", asList(2, 3)), Filter.inArray("a", asList(0, 1, 4))));
-    checkOnlineAndOfflineResultsMatch(query4);
-
+    // Two IN operations on the same field with disjunction.
     // a IN [0,3] || a IN [0,2] should union them (similar to: a IN [0,2,3]).
-    Query query5 =
+    Query query2 =
         collection.where(
             Filter.or(Filter.inArray("a", asList(0, 3)), Filter.inArray("a", asList(0, 2))));
-    checkOnlineAndOfflineResultsMatch(query5, "doc3", "doc6");
-
-    // Nested composite filter on the same field.
-    Query query6 =
-        collection.where(
-            Filter.and(
-                Filter.inArray("a", asList(1, 3)),
-                Filter.or(
-                    Filter.inArray("a", asList(0, 2)),
-                    Filter.and(
-                        Filter.greaterThanOrEqualTo("b", 1), Filter.inArray("a", asList(1, 3))))));
-    checkOnlineAndOfflineResultsMatch(query6, "doc3", "doc4");
-
-    // Nested composite filter on different fields.
-    Query query7 =
-        collection.where(
-            Filter.and(
-                Filter.inArray("b", asList(0, 3)),
-                Filter.or(
-                    Filter.inArray("b", asList(1)),
-                    Filter.and(
-                        Filter.inArray("b", asList(2, 3)), Filter.inArray("a", asList(1, 3))))));
-    checkOnlineAndOfflineResultsMatch(query7, "doc4");
+    checkOnlineAndOfflineResultsMatch(query2, "doc3", "doc6");
   }
 
-  @Ignore
   @Test
   public void testUsingInWithArrayContainsAny() {
+    // TODO(orquery): Enable this test against production when possible.
+    assumeTrue(
+        "Skip this test if running against production because it's not yet supported.",
+        isRunningAgainstEmulator());
     Map<String, Map<String, Object>> testDocs =
         map(
             "doc1", map("a", 1, "b", asList(0)),
@@ -1422,26 +1416,12 @@ public class QueryTest {
 
     Query query2 =
         collection.where(
-            Filter.and(
-                Filter.inArray("a", asList(2, 3)), Filter.arrayContainsAny("b", asList(0, 7))));
-    checkOnlineAndOfflineResultsMatch(query2, "doc3");
-
-    Query query3 =
-        collection.where(
             Filter.or(
                 Filter.and(Filter.inArray("a", asList(2, 3)), Filter.equalTo("c", 10)),
                 Filter.arrayContainsAny("b", asList(0, 7))));
-    checkOnlineAndOfflineResultsMatch(query3, "doc1", "doc3", "doc4");
-
-    Query query4 =
-        collection.where(
-            Filter.and(
-                Filter.inArray("a", asList(2, 3)),
-                Filter.or(Filter.arrayContainsAny("b", asList(0, 7)), Filter.equalTo("c", 20))));
-    checkOnlineAndOfflineResultsMatch(query4, "doc3", "doc6");
+    checkOnlineAndOfflineResultsMatch(query2, "doc1", "doc3", "doc4");
   }
 
-  @Ignore
   @Test
   public void testUsingInWithArrayContains() {
     Map<String, Map<String, Object>> testDocs =
@@ -1479,9 +1459,13 @@ public class QueryTest {
     checkOnlineAndOfflineResultsMatch(query4, "doc3");
   }
 
-  @Ignore
   @Test
   public void testOrderByEquality() {
+    // TODO(orquery): Enable this test against production when possible.
+    assumeTrue(
+        "Skip this test if running against production because order-by-equality is "
+            + "not supported yet.",
+        isRunningAgainstEmulator());
     Map<String, Map<String, Object>> testDocs =
         map(
             "doc1", map("a", 1, "b", asList(0)),