+++ firebase-installations/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Unreleased
+
+# 17.1.3
 *   [changed] Internal changes to improve startup time
 
 # 17.1.2