+++ firebase-perf/CHANGELOG.md
@@ -1,9 +1,20 @@
 # Unreleased
+* [changed] Updated javalite, protoc, protobufjavautil to 3.21.11.
+
+# 20.3.1
+* [changed] Migrated [perfmon] to use standard Firebase executors.
+
+
+## Kotlin
+The Kotlin extensions library transitively includes the updated
+`firebase-performance` library. The Kotlin extensions library has no additional
+updates.
 
 # 20.3.0
 * [fixed] Fixed a `NullPointerException` crash when instrumenting screen
   traces on Android 7, 8, and 9.
   (#4146)
+
 
 ## Kotlin
 The Kotlin extensions library transitively includes the updated
@@ -322,4 +333,3 @@ updates.
 
 # 16.1.0
 * [fixed] Fixed a `SecurityException` crash on certain devices that do not have Google Play Services on them.
-